organised code into folders, and added some more pages
