+++ new/src/about/AboutPage.tsx
@@ -1,0 +1,10 @@
+interface AboutPageProps {
+}
+
+const AboutPage = (props: AboutPageProps) => {
+    return(
+        <div><h1>About Page!!</h1></div>
+    )
+}
+
+export default AboutPage;
