+++ new/src/common/Icon.tsx
@@ -1,4 +1,4 @@
-import { styled } from './stitches.config';
+import { styled } from '../stitches.config';
 
 interface IconProps {
     name: string
