+++ new/src/common/Nav.tsx
@@ -1,4 +1,4 @@
-import { styled } from './stitches.config';
+import { styled } from '../stitches.config';
 
 const Nav = styled('nav', {
     display: 'flex',
