+++ new/src/common/Toggle.tsx
@@ -1,4 +1,4 @@
-import { styled } from './stitches.config';
+import { styled } from '../stitches.config';
 
 const Track = styled('div', {
     display: 'inline-flex',
