+++ new/src/header/Header.tsx
@@ -1,0 +1,10 @@
+interface HeaderProps {
+}
+
+const Header = (props: HeaderProps) => {
+    return(
+        <div><h1>Header!!</h1></div>
+    )
+}
+
+export default Header;
