+++ new/src/readings/Reading.tsx
@@ -1,4 +1,4 @@
-import { styled } from './stitches.config';
+import { styled } from '../stitches.config';
 
 const StyledReading = styled('li', {})
