+++ new/src/readings/Readings.tsx
@@ -1,4 +1,4 @@
-import { styled } from './stitches.config';
+import { styled } from '../stitches.config';
 import Reading from './Reading'
 
 const StyledReadings = styled('ul', {
