+++ new/src/settings/SettingsPage.tsx
@@ -1,0 +1,10 @@
+interface SettingsProps {
+}
+
+const SettingsPage = (props: SettingsProps) => {
+    return(
+        <div><h1>Settings!!</h1></div>
+    )
+}
+
+export default SettingsPage;
